ghc-dt/requirements.txt
langgraph
langchain-openai
langchain-community
faiss-cpu
pypdf
docx2txt
streamlit
sentence-transformers
